+++ crates/ryo3-jiff/src/api.rs
@@ -52,6 +52,7 @@ pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(date, m)?)?;
     m.add_function(wrap_pyfunction!(time, m)?)?;
     m.add_function(wrap_pyfunction!(datetime, m)?)?;
+    m.add_function(wrap_pyfunction!(zoned, m)?)?;
     m.add_function(wrap_pyfunction!(offset, m)?)?;
     m.add_function(wrap_pyfunction!(timespan, m)?)?;